Initial Commit
Client works, but the methods should probably be in a separate thread
in case they block on IO.  It may be work evaluating whether or not
a port or C node would fit better due to the blocking.

Passing source and dest IP addresses are not supported, nor are
interactions or callbacks.